REFACTOR: Rotation changes
I can’t find a good to way to use height/width ratios to always set things to a “perfect” standard, but we can get reasonably close if we simplify it to just checking whether the new CGSize is portrait or landscape.

So now if you rotate, the constraint will reset itself to a better ratio, the beginning constraint that we use to pull on the table view will reset itself to a better ratio, and nothing else changes. The user shouldn’t notice much difference at all when they rotate.