BUG FIX: Offset translation for scrolling the table view
BEFORE:

The front slideable controller is in the up position and scrolled all the way up on the table view (i.e. to the bottom of the table)

Lift finger up and stabilize. Table view scrolled to the bottom, map dimmed, etc.

3a. Swipe gesture down - this should be (and is) picked up by the table view until the table view scrolls all the way to the top…

3b. …whereupon the gesture recognizer takes it over and starts to slide the front view down.

HOWEVER, the constraint was being set continuously based 100% on gesture-translation.y, but without regard to the fact that much of that translation.y was taken up by the table view scrolling back to the top. This makes the front slideable view "jump" downward by however much translation.y was necessary to get the table view back to the top.
AFTER:

In gestureRecognized, get a “snapshot” of the tableViewScrollOffset. This equals however much translation.y it took to scroll the table view all the way back to the top.

Send that snapshot to adjustConstraints to use as an offset for setting the slideyTopConstraint.constant.

Ultimately, this means that the front slideable view won’t “jump” downwards if you scroll the table view with a gesture and continue that gesture, without interruption, to start sliding the front slideable view itself downwards.

This now mimics the behavior in Apple Maps.